scenario: Split into submodules
